+++ weblate/trans/models/comment.py
@@ -20,8 +20,6 @@
 
 from __future__ import unicode_literals
 
-import re
-
 from django.conf import settings
 from django.db import models
 from django.utils.encoding import python_2_unicode_compatible
@@ -29,8 +27,6 @@ from django.utils.encoding import python_2_unicode_compatible
 from weblate.trans.mixins import UserDisplayMixin
 from weblate.trans.models.change import Change
 from weblate.utils.unitdata import UnitData
-
-MENTIONS_RE = re.compile(r'@([\w.@+-]+)\b', re.UNICODE)
 
 
 class CommentManager(models.Manager):
@@ -102,12 +98,3 @@ class Comment(UnitData, UserDisplayMixin):
         return 'comment for {0} by {1}'.format(
             self.content_hash, self.user.username if self.user else 'unknown'
         )
-
-    def get_mentions(self):
-        from weblate.auth.models import User
-
-        for match in MENTIONS_RE.findall(self.comment):
-            try:
-                yield User.objects.get(username=match)
-            except User.DoesNotExist:
-                continue